+++ NEWS.md
@@ -1,5 +1,7 @@
 # officer.pptx 0.1.1
 
+* `add_slide`: gains dynamic dots (passed on to `phs_with`) and arg `.phs_with` to create slide and fill placeholders in one step.
+* `phs_with` gains dynamic dots.
 * `xml_add_image_at_shape`: fix vectorization. Only works for non placeholder shapes.
 
 # officer.pptx 0.1.0